fix : 전체 채팅방이 아닌, 본인 채팅방 전체 검색 | (리팩토링) 채팅방 입장 / 채팅
